Use crossterm and ditch mio
Crossterm doesn't really give access to something I can easily add to
mio to determine if a new event is available, so just use a simple
multi-threaded approach.

+++ cargo-espmonitor/src/main.rs
@@ -133,7 +133,7 @@ fn parse_args(args: Vec<OsString>) -> Result<Option<CargoAppArgs>, Box<dyn Error
                     chip,
                     framework,
                     reset: args.contains("--reset") || !args.contains("--no-reset"),
-                    speed: args.opt_value_from_fn("--speed", |s| s.parse::<u32>())?,
+                    speed: args.opt_value_from_fn("--speed", |s| s.parse::<usize>())?,
                     bin: Some(bin.as_os_str().to_os_string()),
                     serial: args.free_from_str()?,
                 }
